+++ modules/eks/aws/lb.tf
@@ -90,13 +90,17 @@ module "nlb_private" {
 }
 
 resource "aws_route53_record" "wildcard" {
-  count = var.create_public_nlb || var.create_private_nlb ? 1 : 0
+  count = var.base_domain != null && (var.create_public_nlb || var.create_private_nlb) ? 1 : 0
 
-  zone_id = data.aws_route53_zone.this.id
+  zone_id = data.aws_route53_zone.this.0.id
   name    = format("*.apps.%s", var.cluster_name)
   type    = "CNAME"
   ttl     = "300"
   records = [
     var.create_public_nlb ? module.nlb.lb_dns_name : module.nlb_private.lb_dns_name,
   ]
+}
+
+data "dns_a_record_set" "nlb" {
+  host = module.nlb.lb_dns_name
 }