+++ modules/k3s/docker/variables.tf
@@ -1,7 +1,7 @@
 variable "k3s_version" {
   description = "The K3s version to use"
   type        = string
-  default     = "v1.20.7-k3s1"
+  default     = "v1.20.10-k3s1"
 }
 
 variable "server_ports" {
@@ -20,4 +20,22 @@ variable "cluster_endpoint" {
   description = "The api endpoint, when empty it's the container's IP."
   type        = string
   default     = null
+}
+
+variable "worker_groups" {
+  description = "A map defining worker group configurations"
+
+  type = map(object({
+    node_count  = number
+    node_labels = list(string)
+    node_taints = list(string)
+  }))
+
+  default = {
+    "default" = {
+      node_count  = 2
+      node_labels = []
+      node_taints = []
+    }
+  }
 }