+++ tests/sks-exoscale/main.tf
@@ -1,0 +1,9 @@
+module "cluster" {
+  source = "../../modules/sks/exoscale"
+
+  cluster_name = var.cluster_name
+  zone         = "de-fra-1"
+
+  repo_url        = var.repo_url
+  target_revision = var.target_revision
+}